+++ ktor-features/pom.xml
@@ -19,6 +19,7 @@
         <module>ktor-html-builder</module>
         <module>ktor-auth</module>
         <module>ktor-auth-ldap</module>
+        <module>ktor-jwt</module>
     </modules>
 
     <dependencies>
