+++ ktor-samples/ktor-samples-auth/resources/application.conf
@@ -7,8 +7,3 @@ ktor {
         modules = [ org.jetbrains.ktor.samples.auth.JwtAuthApplicationKt.JWTAuthApplication ]
     }
 }
-
-jwt {
-    domain = "https://jwt-provider-domain/"
-    audience = "jwt-audience"
-}
