+++ ktor-samples/ktor-samples-jwt/resources/application.conf
@@ -1,0 +1,14 @@
+ktor {
+    deployment {
+        port = 8080
+    }
+
+    application {
+        modules = [ org.jetbrains.ktor.samples.jwt.JwtApplicationKt.JwtApplication ]
+    }
+}
+
+jwt {
+    domain = "https://jwt-provider-domain/"
+    audience = "jwt-audience"
+}
