+++ ktor-samples/ktor-samples-jwt/web/META-INF/context.xml
@@ -1,0 +1,3 @@
+<Context allowCasualMultipartParsing="true">
+    <Parameter name="org.jetbrains.ktor.config" value="application.conf"/>
+</Context>
