+++ ktor-samples/ktor-samples-jwt/web/WEB-INF/web.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="ISO-8859-1" ?>
+
+<web-app xmlns="http://java.sun.com/xml/ns/javaee"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
+         version="3.0">
+    <servlet>
+        <display-name>Ktor JWT Sample</display-name>
+        <servlet-name>KtorJwtSample</servlet-name>
+        <servlet-class>org.jetbrains.ktor.servlet.ServletApplicationHost</servlet-class>
+        <multipart-config>
+            <!-- 100mb max file upload -->
+            <max-file-size>304857600</max-file-size>
+            <max-request-size>304857600</max-request-size>
+            <file-size-threshold>0</file-size-threshold>
+        </multipart-config>
+    </servlet>
+
+    <servlet-mapping>
+        <servlet-name>KtorJwtSample</servlet-name>
+        <url-pattern>/</url-pattern>
+    </servlet-mapping>
+
+</web-app>
