+++ ktor-samples/pom.xml
@@ -18,6 +18,7 @@
         <module>ktor-samples-async</module>
         <module>ktor-samples-testable</module>
         <module>ktor-samples-json</module>
+        <module>ktor-samples-jwt</module>
         <module>ktor-samples-post</module>
         <module>ktor-samples-embedded</module>
         <module>ktor-samples-httpbin</module>
